made change in info card
